requirements.txt
azure-functions
pandas==2.1.4
markdown==3.5.1
xhtml2pdf==0.2.15
bleach==6.1.0
pytest==7.4.3
requests==2.31.0
playwright==1.57.0
dbfread==2.0.7
chardet==5.2.0

